refactor(lessons): ChapterEditForm sous 300 lignes (H5)
Éclatement complet (pattern G1) : 509 → 225 lignes.
- ChapterContentField.vue (168 l) : éditeurs de contenu par type
  (texte/vidéo/lien/upload) + CSS verbatim.
- ChapterQuizField.vue (168 l) : bloc quiz inline + CSS verbatim.
- Tests de montage pour les deux sous-composants (8 cas).
Parité : mêmes classes/textes, CSS déplacé verbatim, two-way par référence
conservé. Aucune dette pré-existante modifiée.
